tests

+++ Java/lab_5/src/main/java/Task40/Arithmetic.java
@@ -2,8 +2,8 @@ package Task40;
 
 public interface Arithmetic {
 
-    void add(double number);
-    void sub(double number);
-    void mult(double number);
-    void div(double number);
+    void add(int number);
+    void sub(int number);
+    void mult(int number);
+    void div(int number);
 }
